Updating documentation with instruction on using SED charts.
Documention updated with a section that describes how to use SED charts
to test application can bind to a service endpoint definition. This
will be useful for developer that want to make sure they implement
the service binding spec even when they deploy their service manually
or via helm chart.